chore(sync-service): deploy
