+++ apps/iframe/src/listeners/atoms.ts
@@ -2,7 +2,7 @@ import { Msgs } from "@happy.tech/wallet-common"
 import { getDefaultStore } from "jotai/vanilla"
 import { http, createPublicClient } from "viem"
 import { mainnet } from "viem/chains"
-import { permissionsMapLegend } from "#src/state/permissions.ts"
+import { permissionsMapLegend } from "#src/state/permissions/observable"
 import { appMessageBus } from "../services/eventBus"
 import { authStateAtom } from "../state/authState"
 import { userAtom } from "../state/user"
