+++ apps/iframe/src/state/permissions/index.ts
@@ -1,177 +1,14 @@
 import type { Address } from "@happy.tech/common"
 import { permissionsLogger } from "#src/utils/logger"
-import { observable } from "@legendapp/state"
-import { ObservablePersistLocalStorage } from "@legendapp/state/persist-plugins/local-storage"
-import { syncedCrud } from "@legendapp/state/sync-plugins/crud"
 import { PermissionName } from "#src/constants/permissions"
-import { type AppURL, getWalletURL, isApp, isStandaloneWallet } from "../utils/appURL"
-import { checkIfCaveatsMatch } from "../utils/checkIfCaveatsMatch"
-import { emitUserUpdate } from "../utils/emitUserUpdate"
-import { revokedSessionKeys } from "./interfaceState"
-import { getUser } from "./user"
+import { type AppURL, getWalletURL, isApp, isStandaloneWallet } from "#src/utils/appURL"
+import { checkIfCaveatsMatch } from "#src/utils/checkIfCaveatsMatch"
+import { emitUserUpdate } from "#src/utils/emitUserUpdate"
+import { revokedSessionKeys } from "#src/state/interfaceState"
+import { getUser } from "#src/state/user"
+import { permissionsMapLegend } from "./observable"
+import type { AppPermissions, WalletPermission, WalletPermissionCaveat, PermissionsRequest } from "./types"
 import type { HappyUser } from "@happy.tech/wallet-common"
-import { deploymentVar } from "#src/env.ts"
-
-// In EIP-2255, permissions define whether an app can make certain EIP-1193 requests to the wallets.
-// These permissions are scoped per app and per account.
-//
-// The system is not widely adopted and mostly wallet only handles the `eth_accounts` permission,
-// which defines whether an app can get the user's account(s) and subsequently make other requests
-// (some of which will require confirmations, like `eth_sendTransaction`, some of which won't like
-// `eth_call`).
-//
-// Like other wallets, we only handle the `eth_accounts` permission, but we support processing
-// all incoming permission requests.
-//
-// References:
-// https://eips.ethereum.org/EIPS/eip-2255
-
-/**
- * Maps an user + app pair to a {@link AppPermissions}, which is the set of permissions
- * for that user on that app.
- */
-export type PermissionsMap = Record<Address, Record<AppURL, AppPermissions>>
-
-/**
- * Maps permissions names to permission objects.
- * EIP-2255 specifies that permissions names must be EIP-1193 request names (e.g. `eth_accounts`).
- * However, we type this as a string in case we want to extend the permission system to other
- * names that do not map to a request (or are custom requests).
- */
-export type AppPermissions = Record<string, WalletPermission>
-
-/**
- * Permission object for a specific permission.
- *
- * This type is copied from Viem (eip1193.ts)
- */
-export type WalletPermission = {
-    type: "WalletPermissions"
-    // The user to which the permission is granted.
-    user: Address
-    // The app to which the permission is granted.
-    invoker: AppURL
-    // This is the EIP-1193 request that this permission is mapped to.
-    parentCapability: PermissionName | (string & {})
-    caveats: WalletPermissionCaveat[]
-    date: number
-    // Not in the EIP, but Viem wants this.
-    id: string
-    updatedAt: number
-    createdAt: number
-    deleted: boolean
-}
-
-/**
- * A caveat is a specific specific restrictions applied to the permitted request.
- */
-type WalletPermissionCaveat = {
-    type: string
-    value: unknown
-}
-
-/**
- * A request for one or more permissions.
- */
-export type PermissionRequestObject = {
-    [requestName: string]: { [caveatName: string]: unknown }
-}
-
-/**
- * A refinement of {@link PermissionRequestObject} for requesting session keys.
- */
-export type SessionKeyRequest = {
-    [PermissionName.SessionKey]: { target: Address }
-}
-
-/**
- * A permissions specifier, which can be either a single EIP-1193 request name, or a {@link
- * PermissionRequestObject}.
- */
-export type PermissionsRequest = string | PermissionRequestObject
-
-
-const SYNC_SERVICE_URL = deploymentVar("VITE_SYNC_SERVICE_URL")
- 
-export const permissionsMapLegend = observable(
-    syncedCrud({
-        list: async ({ lastSync }) => {
-            const user = getUser()
-            if (!user) return []
-
-            const response = await fetch(
-                `${SYNC_SERVICE_URL}/api/v1/settings/list?type=WalletPermissions&user=${user.address}${lastSync ? `&lastUpdated=${lastSync}` : ""}`,
-            )
-            const data = await response.json()
-
-            return data.data as WalletPermission[]
-        },
-        create: async (data: WalletPermission) => {
-            const response = await fetch(`${SYNC_SERVICE_URL}/api/v1/settings/create`, {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify(data),
-            })
-            await response.json()
-        },
-        update: async (data: WalletPermission) => {
-            const user = getUser()
-            if (!user) return
-
-            const response = await fetch(`${SYNC_SERVICE_URL}/api/v1/settings/update`, {
-                method: "PUT",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({
-                    ...data,
-                    type: "WalletPermissions",
-                    user: user.address,
-                }),
-            })
-            await response.json()
-        },
-        subscribe: ({ refresh }) => {
-            const user = getUser()
-            if (!user) return () => {}
-
-            console.log("Subscribing to updates for user", user.address)
-
-            const eventSource = new EventSource(`${SYNC_SERVICE_URL}/api/v1/settings/subscribe?user=${user.address}`)
-            eventSource.addEventListener("config.changed", (event) => {
-                const data = JSON.parse(event.data)
-                console.log("Received update", data)
-                refresh()
-            })
-
-            return () => eventSource.close()
-        },
-        delete: async ({ id }) => {
-            const response = await fetch(`${SYNC_SERVICE_URL}/api/v1/settings/delete`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ id }),
-            })
-            await response.json()
-        },
-        persist: {
-            plugin: ObservablePersistLocalStorage,
-            name: "config-legend",
-            retrySync: true, // Retry sync after reload
-        },
-        initial: {},
-        fieldCreatedAt: "createdAt",
-        fieldUpdatedAt: "updatedAt",
-        fieldDeleted: "deleted",
-        changesSince: "last-sync",
-        updatePartial: true,
-    }),
-)
-
 
 // === GET ALL PERMISSIONS =======================================================================================
