+++ apps/iframe/src/state/permissions/permissions.spec.ts
@@ -10,7 +10,7 @@ import {
     revokePermissions,
 } from "#src/state/permissions"
 import { disablePermissionWarnings } from "#src/testing/utils"
-import { setUser } from "../state/user"
+import { setUser } from "#src/state/user"
 
 const { appURL, walletURL, appURLMock } = await vi //
     .hoisted(async () => await import("#src/testing/cross_origin.mocks"))
