+++ apps/sync-service/src/env.ts
@@ -1,7 +1,7 @@
 import { z } from "zod"
 
 const envSchema = z.object({
-    NODE_ENV: z.enum(["development", "production"]),
+    NODE_ENV: z.enum(["development", "production", "staging"]),
     APP_PORT: z.string().transform((s) => Number(s)),
     SETTINGS_DB_URL: z.string().trim(),
     LOG_LEVEL: z.preprocess(
