+++ apps/sync-service/src/handlers/subscribe/subscribe.ts
@@ -6,8 +6,6 @@ import type { SubscribeInput } from "./types"
 export async function subscribe(input: SubscribeInput, stream: SSEStreamingApi) {
     const { promise, reject } = promiseWithResolvers<void>()
 
-    console.log("Subscribing to updates for user", input.user)
-
     stream.onAbort(() => {
         reject(undefined)
     })
